Hub Authorizer Lambda Function

+++ hub_api_deployment/prepare_lambda.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+TMP_DIR="/tmp"
+DEFAULT_FILENAME="lambda_function.py"
+LAMBDA_HANDLERS_DIR="../lambda_handlers"
+
 function usage() {
     cat <<EOF
     Usage: $0 [options]
@@ -33,7 +37,12 @@ do
 done
 
 if [[ -n "$FUNCTION_NAME" ]]; then
+    FUNCTION_NAME=${FUNCTION_NAME//.py}
     echo "Got Lambda function name: ${FUNCTION_NAME}"
+    if [[ ! -f  ${LAMBDA_HANDLERS_DIR}/${FUNCTION_NAME}.py ]]; then
+        echo "File  ${LAMBDA_HANDLERS_DIR}/${FUNCTION_NAME}.py doesn't exist! Exiting!"
+        exit 1
+    fi
 else
     echo "No function name passed. Exiting!"
     usage
@@ -55,13 +64,15 @@ function cleanup() {
 
 cleanup
 
-DEFAULT_FILENAME="lambda_function.py"
-LAMBDA_HANDLERS_DIR="../lambda_handlers"
 mkdir -p ${PACKAGE_DIR}
-pip install --target ./${PACKAGE_DIR} -r ${LAMBDA_HANDLERS_DIR}/req_${FUNCTION_NAME}.txt
-cd ${PACKAGE_DIR} && zip -r9 ${OLDPWD}/${FUNCTION_NAME}.zip . && cd -
 
-cp ${LAMBDA_HANDLERS_DIR}/${FUNCTION_NAME}.py /tmp/${DEFAULT_FILENAME}
-cd /tmp && zip -g ${OLDPWD}/${FUNCTION_NAME}.zip ${DEFAULT_FILENAME} && cd -
+# Build package if req_function_name.txt file exists in lambda_handlers directory
+if [[ -f ${LAMBDA_HANDLERS_DIR}/req_${FUNCTION_NAME}.txt ]]; then
+    pip install --target ./${PACKAGE_DIR} -r ${LAMBDA_HANDLERS_DIR}/req_${FUNCTION_NAME}.txt
+    cd ${PACKAGE_DIR} && zip -r9 ${OLDPWD}/${FUNCTION_NAME}.zip . && cd -
+fi
+
+cp ${LAMBDA_HANDLERS_DIR}/${FUNCTION_NAME}.py ${TMP_DIR}/${DEFAULT_FILENAME}
+cd ${TMP_DIR} && zip -g ${OLDPWD}/${FUNCTION_NAME}.zip ${DEFAULT_FILENAME} && cd -
 
 # cleanup
